+++ pkg/api/v1/clients/configmap/resource_client.go
@@ -5,8 +5,8 @@ import (
 	"sort"
 
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
+	"github.com/solo-io/solo-kit/pkg/api/v1/clients/common"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/cache"
-	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/resources/common"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources"
 	"github.com/solo-io/solo-kit/pkg/errors"
 	apiexts "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
