+++ pkg/api/v1/clients/kubesecret/converters.go
@@ -1,4 +1,4 @@
-package secret
+package kubesecret
 
 import (
 	"context"
