+++ pkg/api/v1/clients/kubesecret/kubesecret_suite_test.go
@@ -1,4 +1,4 @@
-package secret_test
+package kubesecret_test
 
 import (
 	"testing"
