+++ pkg/api/v1/clients/kubesecret/plain_secret_resource_client_test.go
@@ -1,11 +1,10 @@
-package configmap_test
+package kubesecret_test
 
 import (
 	"context"
 	"os"
 	"time"
 
-	"github.com/solo-io/go-utils/kubeutils"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/cache"
 
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
@@ -15,7 +14,8 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	. "github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/resources/configmap"
+	"github.com/solo-io/go-utils/kubeutils"
+	. "github.com/solo-io/solo-kit/pkg/api/v1/clients/kubesecret"
 	"github.com/solo-io/solo-kit/pkg/utils/log"
 	"github.com/solo-io/solo-kit/test/helpers"
 	"github.com/solo-io/solo-kit/test/setup"
@@ -27,7 +27,7 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 )
 
-var _ = Describe("PlainConfigmap", func() {
+var _ = Describe("Kube Secret Client Plain=True", func() {
 	if os.Getenv("RUN_KUBE_TESTS") != "1" {
 		log.Printf("This test creates kubernetes resources and is disabled by default. To enable, set RUN_KUBE_TESTS=1 in your env.")
 		return
@@ -37,7 +37,6 @@ var _ = Describe("PlainConfigmap", func() {
 		cfg       *rest.Config
 		client    *ResourceClient
 		kube      kubernetes.Interface
-		kubeCache cache.KubeCoreCache
 	)
 	BeforeEach(func() {
 		namespace = helpers.RandString(8)
@@ -47,10 +46,9 @@ var _ = Describe("PlainConfigmap", func() {
 		Expect(err).NotTo(HaveOccurred())
 		kube, err = kubernetes.NewForConfig(cfg)
 		Expect(err).NotTo(HaveOccurred())
-		kubeCache, err = cache.NewKubeCoreCache(context.TODO(), kube)
+		kcache, err := cache.NewKubeCoreCache(context.TODO(), kube)
 		Expect(err).NotTo(HaveOccurred())
-		client, err = NewResourceClient(kube, &v1.MockResource{}, kubeCache, true)
-		Expect(err).NotTo(HaveOccurred())
+		client, err = NewResourceClient(kube, &v1.MockResource{}, true, kcache)
 	})
 	AfterEach(func() {
 		setup.TeardownKube(namespace)
@@ -72,7 +70,7 @@ var _ = Describe("PlainConfigmap", func() {
 		_, err = client.Write(input, clients.WriteOpts{})
 		Expect(err).NotTo(HaveOccurred())
 
-		cm, err := kube.CoreV1().ConfigMaps(input.Metadata.Namespace).Get(input.Metadata.Name, metav1.GetOptions{})
+		cm, err := kube.CoreV1().Secrets(input.Metadata.Namespace).Get(input.Metadata.Name, metav1.GetOptions{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(cm.Data).To(HaveKey("data.json"))
 		Expect(string(cm.Data["data.json"])).To(Equal("hello: goodbye"))
@@ -91,8 +89,9 @@ var _ = Describe("PlainConfigmap", func() {
 		_, err = client.Write(input, clients.WriteOpts{})
 		Expect(err).NotTo(HaveOccurred())
 
-		cm, err := kube.CoreV1().ConfigMaps(input.Metadata.Namespace).Get(input.Metadata.Name, metav1.GetOptions{})
+		cm, err := kube.CoreV1().Secrets(input.Metadata.Namespace).Get(input.Metadata.Name, metav1.GetOptions{})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(cm.Data).To(HaveKey("data.json"))
 	})
+	// no string escape
 })
