+++ pkg/api/v1/clients/kubesecret/resource_client.go
@@ -1,4 +1,4 @@
-package secret
+package kubesecret
 
 import (
 	"context"
@@ -6,9 +6,8 @@ import (
 	"reflect"
 	"sort"
 
-	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/resources/common"
-
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
+	"github.com/solo-io/solo-kit/pkg/api/v1/clients/common"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/cache"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources"
 	"github.com/solo-io/solo-kit/pkg/errors"
