+++ pkg/api/v1/clients/kubesecret/resource_client_test.go
@@ -1,26 +1,21 @@
-package configmap_test
+package kubesecret_test
 
 import (
 	"context"
 	"os"
 	"time"
 
+	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
 	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/cache"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
 
-	"github.com/solo-io/go-utils/kubeutils"
-
-	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	v1 "github.com/solo-io/solo-kit/test/mocks/v1"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	. "github.com/solo-io/solo-kit/pkg/api/v1/clients/kube/resources/configmap"
-	"github.com/solo-io/solo-kit/pkg/utils/log"
+	"github.com/solo-io/go-utils/kubeutils"
+	. "github.com/solo-io/solo-kit/pkg/api/v1/clients/kubesecret"
 	"github.com/solo-io/solo-kit/test/helpers"
+	v1 "github.com/solo-io/solo-kit/test/mocks/v1"
 	"github.com/solo-io/solo-kit/test/setup"
 	"github.com/solo-io/solo-kit/test/tests/generic"
 	"k8s.io/client-go/kubernetes"
@@ -31,15 +26,16 @@ import (
 )
 
 var _ = Describe("Base", func() {
-	if os.Getenv("RUN_KUBE_TESTS") != "1" {
-		log.Printf("This test creates kubernetes resources and is disabled by default. To enable, set RUN_KUBE_TESTS=1 in your env.")
-		return
-	}
+	BeforeEach(func() {
+		if os.Getenv("RUN_KUBE_TESTS") != "1" {
+			Skip("This test creates kubernetes resources and is disabled by default. To enable, set RUN_KUBE_TESTS=1 in your env.")
+		}
+	})
+
 	var (
 		namespace string
 		cfg       *rest.Config
 		client    *ResourceClient
-		kube      kubernetes.Interface
 		kubeCache cache.KubeCoreCache
 	)
 	BeforeEach(func() {
@@ -48,37 +44,17 @@ var _ = Describe("Base", func() {
 		Expect(err).NotTo(HaveOccurred())
 		cfg, err = kubeutils.GetConfig("", "")
 		Expect(err).NotTo(HaveOccurred())
-		kube, err = kubernetes.NewForConfig(cfg)
+		kube, err := kubernetes.NewForConfig(cfg)
 		Expect(err).NotTo(HaveOccurred())
 		kubeCache, err = cache.NewKubeCoreCache(context.TODO(), kube)
 		Expect(err).NotTo(HaveOccurred())
-		client, err = NewResourceClient(kube, &v1.MockResource{}, kubeCache, false)
-		Expect(err).NotTo(HaveOccurred())
+		client, err = NewResourceClient(kube, &v1.MockResource{}, false, kubeCache)
 	})
 	AfterEach(func() {
 		setup.TeardownKube(namespace)
 	})
 	It("CRUDs resources", func() {
 		generic.TestCrudClient(namespace, client, time.Minute)
-	})
-	It("uses json keys when serializing", func() {
-		foo := "test-data-keys"
-		input := v1.NewMockResource(namespace, foo)
-		data := "hello: goodbye"
-		input.Data = data
-		labels := map[string]string{"pick": "me"}
-		input.Metadata.Labels = labels
-
-		err := client.Register()
-		Expect(err).NotTo(HaveOccurred())
-
-		_, err = client.Write(input, clients.WriteOpts{})
-		Expect(err).NotTo(HaveOccurred())
-
-		cm, err := kube.CoreV1().ConfigMaps(input.Metadata.Namespace).Get(input.Metadata.Name, metav1.GetOptions{})
-		Expect(err).NotTo(HaveOccurred())
-		Expect(cm.Data).To(HaveKey("data.json"))
-		Expect(cm.Data["data.json"]).To(ContainSubstring("'hello: goodbye'"))
 	})
 
 	Context("multiple namespaces", func() {
@@ -101,7 +77,9 @@ var _ = Describe("Base", func() {
 			clientset, err := kubernetes.NewForConfig(cfg)
 			Expect(err).NotTo(HaveOccurred())
 
-			client, err = NewResourceClient(clientset, &v1.MockResource{}, kubeCache, false)
+			kubeCache, err = cache.NewKubeCoreCache(context.TODO(), clientset)
+			Expect(err).NotTo(HaveOccurred())
+			client, err = NewResourceClient(clientset, &v1.MockResource{}, false, kubeCache)
 			Expect(err).NotTo(HaveOccurred())
 		})
 
@@ -170,7 +148,6 @@ var _ = Describe("Base", func() {
 			for {
 				select {
 				case list = <-w:
-					log.Printf("%v", len(list))
 					timesDrained++
 					if timesDrained > 50 {
 						Fail("drained the watch channel 50 times, something is wrong")
